Upgrade tests
Test environment installed.

+++ shared/src/commonMain/kotlin/ru/z8/louttsev/cheaptripmobile/shared/model/LocationRepository.kt
@@ -34,7 +34,7 @@ class LocationRepository(
     fun searchLocationsByName(
         needle: String,
         type: Type = ALL,
-        limit: Int = 10,
+        limit: Long = 10,
         locale: Locale = currentLocale
     ): List<Location> {
         val params = ParamsBundle().apply {
